requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
flask-migrate
waitress
Werkzeug
pandas
reportlab
python-dotenv
psycopg2
pypyodbc
python_odt_template
python-docx
sqlacodegen-v2
flask_caching
flask_wtf
docxtpl
flask_htmx
bcrypt
openpyxl
flask_excel
dotenv